Adaptive-Learning-Web-App-Using-Behaviour-Based-Difficulty-Adjustment-main/ML/requirements.txt
# Upgrade pip, setuptools, wheel first (Render will handle this in build command)
pip>=25.3
setuptools>=68.0
wheel>=0.40

# Scientific & ML packages compatible with Python 3.10
numpy==1.23.5
scikit-learn==1.3.2
pandas==2.1.4
tensorflow-cpu==2.12.0

# Web & API packages
fastapi==0.101.0
uvicorn==0.23.2
joblib==1.3.2
matplotlib==3.7.2
requests==2.32.0
python-dotenv==1.0.0